Simplify the slip handling.
The slip handling needed one buffer to look for packet delimiters, and another buffer for removing escaped characters. These two functions are now merged into the parser. This has eliminated the need for another copy. This functionallty is currently used in Com RX task, and in the param handler. Both are tested.

Cbuff test is updated to test if it is possible to read more than what is in the buffer.