Support for binding of spectrum remote receivers.
Spectrum receivers need a pulse train at the initial startup of the module to enter into bind mode. This is now implemented in the FC and GS. A bind command can be sent from GS to FC and the FC will enter the receiver into bind mode. No power cycling is needed.