Support for more than 32 message types added. Satellite receiver now able to bind.
The event hander subscriptions were stored in a 32 bit value where each bit reprresented a subscription of that message type. This only allowed for 32 message types, and that limit was reached. Now the subscriptions is an array of 32 bit numbers, so there is uint32_t_MAX number of messages allowed in the system now. The limitation is that each handler can not subscribe to all messages at the same time. Only 16 subscriptions per handler is allowed currently.