Fix paramHandler.
ParamHandler had some bugs, and it also mixed the paramHandler master and normal param handlers in a way that made the code very messy. ParamHandler master is now broken out to its own module, paramMaster and does not have its own parameters.

The paramMaster must currently be instantiated in a thread that does not have any parameters as it is using some of the same messages as a paramHandler.